+++ src/services/stepService.js
@@ -84,6 +84,9 @@ export const stepService = {
     * @throws {Error} If the API request fails (e.g., validation errors, 404 Not Found, 403 Forbidden).
     */
     updateStep: async (stepId, stepData) => {
+        if (typeof stepId !== 'number' || isNaN(stepId) || stepId < 1) {
+            return Promise.reject(new Error("step id not valid."));
+        }
         const response = await api.put(`${BASE_URL}/${stepId}`, stepData);
         return response.data;
     },
@@ -97,6 +100,9 @@ export const stepService = {
     * @throws {Error} If the API request fails (e.g., 404 Not Found, 403 Forbidden).
     */
     deleteStep: async (stepId) => {
+        if (typeof stepId !== 'number' || isNaN(stepId) || stepId < 1) {
+            return Promise.reject(new Error("step id not valid."));
+        }
         await api.delete(`${BASE_URL}/${stepId}`);
         return true;
     },
@@ -124,6 +130,9 @@ export const stepService = {
     * @throws {Error} If the API request fails (e.g., 404 Not Found, 403 Forbidden).
     */
     getStepByIdAdmin: async (stepId) => {
+        if (typeof stepId !== 'number' || isNaN(stepId) || stepId < 1) {
+            return Promise.reject(new Error("step id not valid."));
+        }
         const response = await api.get(`${ADMIN_BASE_URL}/${stepId}/`);
         return response.data;
     },
@@ -140,6 +149,9 @@ export const stepService = {
     * @throws {Error} If the API request fails (e.g., validation errors, 404 Not Found, 403 Forbidden).
     */
     updateStepAdmin: async (stepId, stepData) => {
+        if (typeof stepId !== 'number' || isNaN(stepId) || stepId < 1) {
+            return Promise.reject(new Error("step id not valid."));
+        };
         const response = await api.patch(`${ADMIN_BASE_URL}/${stepId}/`, stepData);
         return response.data;
     },
@@ -152,7 +164,10 @@ export const stepService = {
     * @returns {Promise<boolean>} A promise that resolves with `true` if the step is successfully deleted.
     * @throws {Error} If the API request fails (e.g., 404 Not Found, 403 Forbidden).
     */
-    deleteStepAdmin: async (stepId) => {
+    deleteStepAdmin: async (stepId) => {        
+        if (typeof stepId !== 'number' || isNaN(stepId) || stepId < 1) {
+            return Promise.reject(new Error("step id not valid."));
+        };
         await api.delete(`${ADMIN_BASE_URL}/${stepId}/`);
         return true;
     }
